use page.title to determine active navbar item
page.slug adds a dependency on the user to create a custom front-matter
item. The title and subtitle of the page are required parameters that
cannot be avoided. Since the check is a hard-coded comparison, the
right-hand-side might as well use the titles of the page.

There is an edge case where changing the case of the title on the page
may break the navigation marker since the title is case-sensitive.

The proper solution would be to use the 'slug' of the page which is the
lower-cased, non-space string of the title of the page. However,
computing the slug within the liquid conditional does not work.