make stripe-social functional
The corresponding CSS and JS files had to be enabled in
`_layouts/base.html`. However, while the buttons work, they are not
visually similar to the original because of the dependency on
site-specific styles. There are a couple inline styles applied to the
top-level `<div>` to make the diaspora and mastodon buttons legible.