import CSS from CDN to build with GitHub Pages
GitHub Pages isolates checkout to only the site directory ( `/docs` )
which would have broken symlinks to the fork-awesome CSS which resides
outside of `/docs`. This change is made to stop GitHub Pages complaining
about broken symlinks.

It is ironic to have to import the project's own artificats to be
imported from a CDN; but that is better than having to duplicate the
same files in the repository without symlinks. But this is a temporary
fix to pacify GitHub Pages. This change MUST NOT go into the final
output.

This reverts commit dc118958dc43750372c0a4518fd7135d8f7d52d3.